% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.R
\name{envelope}
\alias{envelope}
\title{envelope simulado}
\usage{
envelope(
  mod1,
  mod2,
  subtitle1 = "Dffits do Modelo 1",
  subtitle2 = "Dffits do Modelo 2"
)
}
\arguments{
\item{model}{2  model 2}
}
\description{
a paz interna resid
}
